refactor: DDD
