+++ src/platform/workspace/auth/WorkspaceAuthGate.test.ts
