+++ src/platform/workspace/auth/WorkspaceAuthGate.vue
