+++ src/platform/workspace/components/PricingTableWorkspace.vue
