+++ src/platform/workspace/components/SubscriptionTransitionPreviewWorkspace.vue
