+++ src/platform/workspace/components/WorkspaceProfilePic.vue
