+++ src/platform/workspace/components/dialogs/CreateWorkspaceDialogContent.vue
