+++ src/platform/workspace/components/dialogs/DeleteWorkspaceDialogContent.vue
