+++ src/platform/workspace/components/dialogs/EditWorkspaceDialogContent.vue
