+++ src/platform/workspace/components/dialogs/InviteMemberDialogContent.vue
