+++ src/platform/workspace/components/dialogs/InviteMemberUpsellDialogContent.vue
