+++ src/platform/workspace/components/dialogs/LeaveWorkspaceDialogContent.vue
