+++ src/platform/workspace/components/dialogs/RemoveMemberDialogContent.vue
