+++ src/platform/workspace/components/dialogs/RevokeInviteDialogContent.vue
