+++ src/platform/workspace/components/dialogs/settings/MembersPanelContent.vue
