+++ src/platform/workspace/composables/useWorkspaceSwitch.ts
