+++ src/platform/workspace/workspaceTypes.ts
