+++ src/services/dialogService.ts
@@ -568,7 +568,7 @@ export const useDialogService = () => {
     workspaceName?: string
   }) {
     const { default: component } =
-      await import('@/components/dialog/content/workspace/DeleteWorkspaceDialogContent.vue')
+      await import('@/platform/workspace/components/dialogs/DeleteWorkspaceDialogContent.vue')
     return dialogStore.showDialog({
       key: 'delete-workspace',
       component,
@@ -581,7 +581,7 @@ export const useDialogService = () => {
     onConfirm?: (name: string) => void | Promise<void>
   ) {
     const { default: component } =
-      await import('@/components/dialog/content/workspace/CreateWorkspaceDialogContent.vue')
+      await import('@/platform/workspace/components/dialogs/CreateWorkspaceDialogContent.vue')
     return dialogStore.showDialog({
       key: 'create-workspace',
       component,
@@ -598,7 +598,7 @@ export const useDialogService = () => {
 
   async function showLeaveWorkspaceDialog() {
     const { default: component } =
-      await import('@/components/dialog/content/workspace/LeaveWorkspaceDialogContent.vue')
+      await import('@/platform/workspace/components/dialogs/LeaveWorkspaceDialogContent.vue')
     return dialogStore.showDialog({
       key: 'leave-workspace',
       component,
@@ -608,7 +608,7 @@ export const useDialogService = () => {
 
   async function showEditWorkspaceDialog() {
     const { default: component } =
-      await import('@/components/dialog/content/workspace/EditWorkspaceDialogContent.vue')
+      await import('@/platform/workspace/components/dialogs/EditWorkspaceDialogContent.vue')
     return dialogStore.showDialog({
       key: 'edit-workspace',
       component,
@@ -624,7 +624,7 @@ export const useDialogService = () => {
 
   async function showRemoveMemberDialog(memberId: string) {
     const { default: component } =
-      await import('@/components/dialog/content/workspace/RemoveMemberDialogContent.vue')
+      await import('@/platform/workspace/components/dialogs/RemoveMemberDialogContent.vue')
     return dialogStore.showDialog({
       key: 'remove-member',
       component,
@@ -635,7 +635,7 @@ export const useDialogService = () => {
 
   async function showInviteMemberDialog() {
     const { default: component } =
-      await import('@/components/dialog/content/workspace/InviteMemberDialogContent.vue')
+      await import('@/platform/workspace/components/dialogs/InviteMemberDialogContent.vue')
     return dialogStore.showDialog({
       key: 'invite-member',
       component,
@@ -651,7 +651,7 @@ export const useDialogService = () => {
 
   async function showInviteMemberUpsellDialog() {
     const { default: component } =
-      await import('@/components/dialog/content/workspace/InviteMemberUpsellDialogContent.vue')
+      await import('@/platform/workspace/components/dialogs/InviteMemberUpsellDialogContent.vue')
     return dialogStore.showDialog({
       key: 'invite-member-upsell',
       component,
@@ -667,7 +667,7 @@ export const useDialogService = () => {
 
   async function showRevokeInviteDialog(inviteId: string) {
     const { default: component } =
-      await import('@/components/dialog/content/workspace/RevokeInviteDialogContent.vue')
+      await import('@/platform/workspace/components/dialogs/RevokeInviteDialogContent.vue')
     return dialogStore.showDialog({
       key: 'revoke-invite',
       component,
