+++ src/views/layouts/LayoutDefault.vue
@@ -9,7 +9,7 @@
 <script setup lang="ts">
 import { useFavicon } from '@vueuse/core'
 
-import WorkspaceAuthGate from '@/components/auth/WorkspaceAuthGate.vue'
+import WorkspaceAuthGate from '@/platform/workspace/auth/WorkspaceAuthGate.vue'
 
 useFavicon('/assets/favicon.ico')
 </script>
